Calendar: Add additional calendar gml  properties
LIBJS: Move parse_simplified_iso8601 to Core::Datetime
Move the parse_simplified_iso8601 so it can be used in other places
Calendar: Use parse_simplified_iso8601 for parsing date property
